+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+
+## [2.5.0] - TBD
 ### Added
 - Add `AccountBalance.getMarketRegistry()` to return the address of MarketRegistry contract.
 - Add `AccountBalance.getMarkPrice(address)` to return the mark price of given market.